LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_anchor_download_allow_downloads_without_user_activation.sub.tentative-expected.txt

CONSOLE MESSAGE: line 1489: Error while parsing the 'sandbox' attribute: 'allow-downloads-without-user-activation' is an invalid sandbox flag.
CONSOLE MESSAGE: Blocked script execution in 'about:srcdoc' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.


FAIL <a download> triggered download in sandbox is allowed by allow-downloads-without-user-activation. undefined is not an object (evaluating 'anchor.href = "support/download_stash.py?token=" + token')
